Place histograms in double, now reproduce nvt results

+++ Saxs/saxsDeviceKernels.cuh
@@ -16,9 +16,9 @@
 #include <fmt/core.h>
 void scatterCalculation(cuFloatComplex *grid_q, cuFloatComplex *grid_oq, float *oc,
                         float *Scatter, int nnx, int nny, int nnz, float qcut);
-__global__ void calculate_histogram(cuFloatComplex *d_array, float *d_histogram, double *nhist, float *oc, int nx, int ny, int nz,
+__global__ void calculate_histogram(cuFloatComplex *d_array, double *d_histogram, double *nhist, float *oc, int nx, int ny, int nz,
                                     float bin_size, float Qcut, int num_bins, float fact);
-__global__ void calculate_histogram(cuFloatComplex *d_array, float *d_histogram, double *nhist, float *oc, int nx, int ny, int nz,
+__global__ void calculate_histogram(cuFloatComplex *d_array, double *d_histogram, double *nhist, float *oc, int nx, int ny, int nz,
                                     float bin_size, float Qcut, int num_bins);
 
 __global__ void modulusKernel(cuFloatComplex *grid_q, float *modX, float *modY, float *modZ,
